添加注册接口V2

+++ partner-tianyancha-client/src/main/java/com/sbm/module/partner/tianyancha/rest/api736/domain/Api736.java
@@ -51,7 +51,7 @@ public class Api736 {
 
 	private Long toTime;
 
-	private Integer legalPersonId;
+	private Long legalPersonId;
 
 	private String sourceFlag;
